splits handlers since it went over max lineno
